user-service/requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
PyJWT==2.9.0
cryptography==42.0.8
